Crate swarm_nl

SwarmNL

A library to build custom networking layers for decentralized applications

SwarmNL is a library designed for P2P networking in distributed systems. It’s lightweight, scalable, and easy to configure, making it perfect for decentralized applications. Powered by libp2p, SwarmNL simplifies networking, so you can focus on building.

Why SwarmNL?

SwarmNL makes buiding a peer-to-peer decentralized and distributed networking stack for your application a breeze. With SwarmNL, you can effortlessly configure nodes, tailor network conditions, and fine-tune behaviors specific to your project’s needs, allowing you to dive into networking without the usual hassle.

Say goodbye to the complexities of networking and hello to simplicity. With SwarmNL, all the hard work is done for you, leaving you to focus on simple configurations and your application logic. It’s that easy!

Have a look at the [swarmnl_guides::guides] to get started.

Features

Node Configuration: SwarmNL provides a simple interface to configure a node and specify parameters to dictate its behaviour. This includes:

Selection and configuration of the transport layers to be supported by the node.

Selection of cryptographic keypairs (ed25519, RSA, secp256k1, ecdsa)

Storage and retrieval of keypair locally.

PeerID and multiaddress generation.

Protocol specification and handlers.

Event handlers for network events and logging.

Example

      //! Using the default node setup configuration and the default network event handler

      // Default config
      let config = BootstrapConfig::default();
      // Default network handler
      let handler = DefaultHandler;
      // Build node or network core
      let node = CoreBuilder::with_config(config, handler)
          .build()
          .await
          .unwrap();

      //! Using a custom node setup configuration and a custom network event handler

      // Custom configuration
      // a. Using config from an `.ini` file
      let config = BootstrapConfig::from_file("bootstrap_config.ini");

      // b. Using config methods
      let mut bootnode = HashMap::new();  // Bootnodes
      let ports = (1509, 2710);  // TCP, UDP ports

      bootnode.insert(
          PeerId::random(),
          "/ip4/x.x.x.x/tcp/1509".to_string()
      );

      let config = BootstrapConfig::new()
          .with_bootnodes(bootnode)
          .with_tcp(ports.0)
          .with_udp(ports.1);

      // Custom event handler
      use swarm_nl::core::EventHandler;

      #[derive(Clone)]
      struct ApplicationState{
          name: String,
          version: i8,
      }

      // Define custom behaviour to respond to network events
      impl EventHandler for AppState {
          fn new_listen_addr(
              &mut self,
              local_peer_id: PeerId,
              listener_id: ListenerId,
              addr: Multiaddr,
          ) {
              // Announce interfaces we're listening on
              println!("Peer id: {}", local_peer_id);
              println!("We're listening on the {}", addr);
          }

          // Echo data recieved from a RPC
          fn rpc_handle_incoming_message(&mut self, data: Vec<Vec<u8>>) -> Vec<Vec<u8>> {
              println!("Recvd incoming RPC: {:?}", data);
              data
          }

          // Handle the incoming gossip message
          fn gossipsub_handle_incoming_message(&mut self, source: PeerId, data: Vec<String>) {
              println!("Recvd incoming gossip: {:?}", data);
          }
      }

      // Define custom event handler
      let state = ApplicationState {
          name: String::from("SwarmNL"),
          version: 0.1
      }

      // Build node or network core
      let node = CoreBuilder::with_config(config, state)
          .build()
          .await
          .unwrap();

Please look at a template .ini file here for configuring a node in the network.

Node Communication: For communication, SwarmNL leverages the powerful capabilities of libp2p. These includes:

The Kadmlia DHT: Developers can use the DHT to store infomation and leverage the capabilities of the DHT to build powerful applications, easily.

A simple RPC mechanism to exchange data quickly between peers.

Gossiping: SwarmNL uses the Gossipsub 1.1 protocol, specified by the libp2p spec.

Example

      //! Communicate with remote nodes using the simple and familiar async-await paradigm.

      // Build node or network core
      let node = CoreBuilder::with_config(config, state)
          .build()
          .await
          .unwrap();

      // Communication interfaces
      // a. Kademlia DHT e.g

      // Prepare an kademlia `store_record` request to send to the network layer
      let (key, value, expiration_time, explicit_peers) = (
          KADEMLIA_TEST_KEY.as_bytes().to_vec(),
          KADEMLIA_TEST_VALUE.as_bytes().to_vec(),
          None,
          None,
      );

      let kad_request = AppData::KademliaStoreRecord {
          key: key.clone(),
          value,
          expiration_time,
          explicit_peers,
      };

      // Send request
      if let Ok(result) = node.query_network(kad_request).await {
          assert_eq!(KademliaStoreRecordSuccess,result);
	    }

      // b. RPC (request-response) e.g

      // Prepare a RPC fetch request
      let fetch_key = vec!["SomeFetchKey".as_bytes().to_vec()];

      let fetch_request = AppData::FetchData {
          keys: fetch_key.clone(),
          peer: node4_peer_id,
      };

      // Get a stream id to track the request
      let stream_id = node.send_to_network(fetch_request).await.unwrap();

      // Poll for the result
      if let Ok(result) = node.recv_from_network(stream_id).await {
          // Here, the request data was simply echoed by the remote peer
          assert_eq!(AppResponse::FetchData(fetch_key), result);
      }

      // c. Gossiping e.g

      // Prepare gossip request
      let gossip_request = AppData::GossipsubBroadcastMessage {
          topic: GOSSIP_NETWORK.to_string(),
          message: vec!["Daniel".to_string(), "Deborah".to_string()],
      };

      if let Ok(result) = node.query_network(gossip_request).await {
          assert_eq!(AppResponse::GossipsubBroadcastSuccess, result);
      }

In Development 👷:

Node failure handling involving reconnection strategies, failover mechanisms etc.

Scaling involving techniques like sharding, data forwarding etc.

IPFS upload and download interfaces.

Again, SwarmNL simplifies networking so you can focus on building that world-changing application of yours! Cheers! 🥂

With ❤️ from Deji and Sacha.

Modules

core

Core data structures and protocol implementations for building a swarm.

prelude 🔒

Types and traits that are used throughout SwarmNL.

setup

Data structures and functions to setup a node and configure it for networking.

testing_guide

A doc-only module explaining how to run core library tests.

util

Utility helper functions for reading from and writing to .ini config files.

Structs

ConnectionId

Connection identifier.

Keypair

Identity keypair of a node.

ListenerId

The ID of a single listener.

Multiaddr

Representation of a Multiaddr.

PeerId

Identifier of a peer of the network.

RsaKeypair

An RSA keypair.

Enums

ConnectedPoint

The endpoint roles associated with an established peer-to-peer connection.

Failure

An outbound ping failure.

KeyType

The type of key a KeyPair is holding.

SwarmNlError

Library error type containing all custom errors that could be encountered.

TcpConfig

TCP setup configuration.

TransportOpts

Supported transport protocols.

Constants

MAX_PORT

Upper bound port range (u16::MAX).

MIN_PORT

Lower bound port range (u16::MIN).

Statics

DEFAULT_IP_ADDRESS

Default IP address when no address is specified.

DEFAULT_KEEP_ALIVE_DURATION

Default amount of time to keep a connection alive.

DEFAULT_NETWORK_ID

Default network id

MIN_NETWORK_ID_LENGTH

This constant sets the shortest acceptable length for a network ID. The network ID identifies a network and ensures it’s distinct from others.

Traits

CustomFrom

An implementation of From<&str> for KeyType to read a key type from a bootstrap config file.

Type Aliases

MultiaddrString

The stringified Multiaddr type.

PeerIdString

The stringified PeerId type.

Port

Port type.

Seconds

Seconds type.

SwarmNlResult

Generic SwarmNl result type.