Module swarm_nl::core::prelude

Modules

gossipsub_cfg 🔒

Module containing important state relating to the Gossipsub protocol

ping_config

Module that contains important data structures to manage Ping operations on the network

Structs

DefaultHandler

Default network event handler

ExecQueue 🔒

Network queue that tracks the execution of application requests in the network layer

NetworkInfo 🔒

Important information to obtain from the CoreBuilder, to properly handle network operations

RpcConfig

The configuration for the RPC protocol.

StreamId

A simple struct used to track requests sent from the application layer to the network layer.

StreamRequestBuffer 🔒

Type that keeps track of the requests from the application layer. This type has a maximum buffer size and will drop subsequent requests when full. It is unlikely to be ever full as the default is usize::MAX except otherwise specified during configuration. It is always good practice to read responses from the internal stream buffer using query_network() or explicitly using recv_from_network

StreamResponseBuffer 🔒

Type that keeps track of the response to the requests from the application layer.

Enums

AppData

Request sent from the application layer to the networking layer.

AppResponse

Response to requests sent from the application to the network layer.

NetworkError

Network error type containing errors encountered during network operations.

Rpc 🔒

Type representing the RPC data structure sent between nodes in the network.

StreamData 🔒

Data exchanged over a stream between the application and network layer

Constants

BOOT_WAIT_TIME 🔒

Time to wait (in seconds) for the node (network layer) to boot.

GOSSIP_MESSAGE_SEPARATOR 🔒

The delimeter that separates the messages to gossip

NETWORK_READ_TIMEOUT

The duration (in seconds) to wait for response from the network layer before timing out.

STREAM_BUFFER_CAPACITY 🔒

The buffer capacity of an mpsc stream.

TASK_SLEEP_DURATION

The time it takes for the task to sleep before it can recheck if an output has been placed in the repsonse buffer (7 seconds).

Traits

EventHandler

The high level trait that provides an interface for the application layer to respond to network events.

Type Aliases

AppResponseResult 🔒

Type that represents the response of the network layer to the application layer’s event handler.

NetworkResult

Type that contains the result of querying the network layer.