Struct swarm_nl::core::Core

pub struct Core<T: EventHandler + Clone + Send + Sync + 'static> {
    keypair: Keypair,
    application_sender: Sender<StreamData>,
    stream_response_buffer: Arc<Mutex<StreamResponseBuffer>>,
    stream_request_buffer: Arc<Mutex<StreamRequestBuffer>>,
    current_stream_id: Arc<Mutex<StreamId>>,
    pub state: T,
}

The core interface for the application layer to interface with the networking layer.

Fields
keypair: Keypairapplication_sender: Sender<StreamData>

The producing end of the stream that sends data to the network layer from the application.

stream_response_buffer: Arc<Mutex<StreamResponseBuffer>>

The consuming end of the stream that recieves data from the network layer. The producing end of the stream that sends data from the network layer to the application. This serves as a buffer for the results of the requests to the network layer. With this, applications can make async requests and fetch their results at a later time without waiting. This is made possible by storing a StreamId for a particular stream request.

stream_request_buffer: Arc<Mutex<StreamRequestBuffer>>

Store a StreamId representing a network request

current_stream_id: Arc<Mutex<StreamId>>

Current stream id. Useful for opening new streams, we just have to bump the number by 1

state: T

The state of the application

Implementations

impl<T: EventHandler + Clone + Send + Sync + 'static> Core<T>

pub fn save_keypair_offline(&self, config_file_path: &str) -> bool

Serialize keypair to protobuf format and write to config file on disk. This could be useful for saving a keypair for future use when going offline.

It returns a boolean to indicate success of operation. Only key types other than RSA can be serialized to protobuf format and only a single keypair can be saved at a time.

pub fn peer_id(&self) -> PeerId

Return the node’s PeerId

pub async fn send_to_network( &mut self, app_request: AppData ) -> Option<StreamId>

Send data to the network layer and recieve a unique StreamId to track the request.

If the internal stream buffer is full, None will be returned.

pub async fn recv_from_network(&mut self, stream_id: StreamId) -> NetworkResult

Explicitly retrieve the reponse to a request sent to the network layer. This function is decoupled from the Core::send_to_network() method so as to prevent blocking until the response is returned.

pub async fn query_network(&mut self, request: AppData) -> NetworkResult

Perform an atomic send and recieve to and from the network layer. This function is atomic and blocks until the result of the request is returned from the network layer.

This function should mostly be used when the result of the request is needed immediately and delay can be condoned. It will still timeout if the delay exceeds the configured period.

If the internal buffer is full, it will return an error.

async fn handle_network_response( receiver: Receiver<StreamData>, network_core: Core<T> )

Handle the responses coming from the network layer. This is usually as a result of a request from the application layer

async fn handle_async_operations( swarm: Swarm<CoreBehaviour>, network_info: NetworkInfo, network_sender: Sender<StreamData>, receiver: Receiver<StreamData>, network_core: Core<T> )

Handle async operations, which basically involved handling two major data sources:

Streams coming from the application layer.

Events generated by (libp2p) network activities.

Important information are sent to the application layer over a (mpsc) stream.