Struct swarm_nl::core::CoreBuilder

pub struct CoreBuilder<T: EventHandler + Clone + Send + Sync + 'static> {

    network_id: StreamProtocol,
    keypair: Keypair,
    tcp_udp_port: (Port, Port),
    boot_nodes: HashMap<PeerIdString, MultiaddrString>,
    blacklist: Blacklist,
    handler: T,
    stream_size: usize,
    ip_address: IpAddr,
    keep_alive_duration: Seconds,
    transport: TransportOpts,
    ping: (Behaviour, PingErrorPolicy),
    kademlia: Behaviour<MemoryStore>,
    identify: Behaviour,
    request_response: Behaviour<Rpc, Rpc>,
    gossipsub: Behaviour,

}

Structure containing necessary data to build Core.

Fields
network_id: StreamProtocolkeypair: Keypairtcp_udp_port: (Port, Port)boot_nodes: HashMap<PeerIdString, MultiaddrString>blacklist: Blacklisthandler: T

the network event handler

stream_size: usize

The size of the stream buffers to use to track application requests to the network layer internally.

ip_address: IpAddrkeep_alive_duration: Seconds

Connection keep-alive duration while idle.

transport: TransportOpts

The transport protocols being used. TODO: This can be a collection in the future to support additive transports.

ping: (Behaviour, PingErrorPolicy)

The Behaviour of the Ping protocol.

kademlia: Behaviour<MemoryStore>

The Behaviour of the Kademlia protocol.

identify: Behaviour

The Behaviour of the Identify protocol.

request_response: Behaviour<Rpc, Rpc>

The Behaviour of the Request-Response protocol. The second field value is the function to handle an incoming request from a peer.

gossipsub: Behaviour

The Behaviour of the GossipSub protocol

Implementations

impl<T: EventHandler + Clone + Send + Sync + 'static> CoreBuilder<T>

pub fn with_config(config: BootstrapConfig, handler: T) -> Self

Return a CoreBuilder struct configured with BootstrapConfig and default values. Here, it is certain that BootstrapConfig contains valid data. A type that implements EventHandler is passed to handle and responde to network events.

pub fn with_network_id(self, protocol: String) -> Self

Explicitly configure the network (protocol) id.

Note that it must be of the format “/protocol-name/version” otherwise it will default to “/swarmnl/1.0”. See: DEFAULT_NETWORK_ID.

pub fn listen_on(self, ip_address: IpAddr) -> Self

Configure the IP address to listen on.

If none is specified, the default value is Ipv4Addr::new(0, 0, 0, 0). See: DEFAULT_IP_ADDRESS.

pub fn with_idle_connection_timeout(self, keep_alive_duration: Seconds) -> Self

Configure how long to keep a connection alive (in seconds) once it is idling.

pub fn with_stream_size(self, size: usize) -> Self

Configure the size of the stream buffers to use to track application requests to the network layer internally. This should be as large an possible to prevent dropping off requests to the network layer. Defaults to usize::MAX

pub fn with_ping(self, config: PingConfig) -> Self

Configure the Ping protocol for the network.

pub fn with_rpc<F>(self, config: RpcConfig) -> Self

Configure the RPC protocol for the network.

pub fn with_kademlia(self, config: Config) -> Self

Configure the Kademlia protocol for the network.

pub fn with_gossipsub(self, config: Config, auth: MessageAuthenticity) -> Self

Configure the Gossipsub protocol for the network

Panics

This function panics if Gossipsub cannot be configured properly

pub fn with_transports(self, transport: TransportOpts) -> Self

Configure the transports to support.

pub fn configure_network_events(self, handler: T) -> Self

Configure a handler to respond to network events.

pub fn network_id(&self) -> String

Return the id of the network

pub async fn build(self) -> SwarmNlResult<Core<T>>

Build the Core data structure.

Handles the configuration of the libp2p Swarm structure and the selected transport protocols, behaviours and node identity for tokio and async-std runtimes. The Swarm is wrapped in the Core construct which serves as the interface to interact with the internal networking layer.