Module swarm_nl::core::tests::layer_communication

Tests for the communication between the layers of the application.

Structs

AppState

Sate of the Application

Constants

GOSSIP_NETWORK

The test network we join for our mesh

ITEST_WAIT_TIME

Time to wait for the other peer to act, during integration tests (in seconds)

KADEMLIA_TEST_KEY

The key to test the Kademlia DHT

KADEMLIA_TEST_VALUE

The value to test the Kademlia DHT

Functions

setup_core_builder_1

setup_node_1

Used to create a detereministic node.

setup_node_2

Used to create a node to peer with node_1.