Module swarm_nl::prelude

Types and traits that are used throughout SwarmNL.

Enums

SwarmNlError

Library error type containing all custom errors that could be encountered.

TcpConfig

TCP setup configuration.

TransportOpts

Supported transport protocols.

Constants

MAX_PORT

Upper bound port range (u16::MAX).

MIN_PORT

Lower bound port range (u16::MIN).

Statics

DEFAULT_IP_ADDRESS

Default IP address when no address is specified.

DEFAULT_KEEP_ALIVE_DURATION

Default amount of time to keep a connection alive.

DEFAULT_NETWORK_ID

Default network id

MIN_NETWORK_ID_LENGTH

This constant sets the shortest acceptable length for a network ID. The network ID identifies a network and ensures it’s distinct from others.

Traits

CustomFrom

An implementation of From<&str> for KeyType to read a key type from a bootstrap config file.

Type Aliases

MultiaddrString

The stringified Multiaddr type.

PeerIdString

The stringified PeerId type.

Port

Port type.

Seconds

Seconds type.

SwarmNlResult

Generic SwarmNl result type.